refactor(users): move extractEmailDomain to server-only email-domain.ts
Security-report follow-up on the tldts dependency:

- Move extractEmailDomain out of lib/utils.ts (imported by many client components) and into a dedicated lib/email-domain.ts marked with 'server-only'. tldts ships a ~3 MB Public Suffix List and lacks 'sideEffects: false', so keeping it out of client bundles requires an explicit server-only barrier rather than relying on tree-shaking.
- Enable allowPrivateDomains: true so hosts on private suffixes like *.vercel.app, *.github.io, *.s3.amazonaws.com resolve to the per-tenant subdomain (e.g. alice.vercel.app) rather than collapsing every tenant to vercel.app. Better signal for admin abuse-pattern grouping.
- Document in the JSDoc that email_domain is for admin grouping only (not a security boundary) and that keeping tldts current keeps the bundled PSL current.
- Move the unit tests alongside the new file and add coverage for the private-suffix behaviour.

Points from the security report not actioned, with rationale: exact pins (caret + lockfile is fine outside a security boundary), input-length caps (email column is already bounded upstream), SSRF/auth controls (N/A — this column does not gate any decision), license review (MIT package embedding MPL-2.0 PSL data is the common pattern and not something this change introduces).
